# apps/grincajg_rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Grincajg API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/organization": {
            "get": {
                "description": "This endpoint returns the details of the organization the current logged in user belongs to.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Get the current user's organization",
                "responses": {
                    "200": {
                        "description": "Organization data successfully returned",
                        "schema": {
                            "$ref": "#/definitions/models.Organization"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not logged in",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "User does not belong to any organization",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/organizations": {
            "post": {
                "description": "This endpoint creates a new organization and assigns the current logged in user as the admin of the organization.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Create a new organization",
                "parameters": [
                    {
                        "description": "Organization creation input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateOrganizationInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Organization successfully created",
                        "schema": {
                            "$ref": "#/definitions/models.OrganizationRecord"
                        }
                    },
                    "400": {
                        "description": "Bas request",
                        "schema": {
                            "$ref": "#/definitions/models.ValidateErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not logged in",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "This endpoint allows a user to log in by providing necessary information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Log in a user",
                "parameters": [
                    {
                        "description": "User login details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignInInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token successfully returned",
                        "schema": {
                            "$ref": "#/definitions/models.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Validation error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway: Something bad happened",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "get": {
                "description": "This endpoint allows a user to log out.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Log out a user",
                "responses": {
                    "200": {
                        "description": "User successfully logged out",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "description": "This endpoint returns the details of the current logged in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get the current user",
                "responses": {
                    "200": {
                        "description": "Current user data successfully returned",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not logged in",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "This endpoint allows a user to sign up by providing necessary information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Sign up a new user",
                "parameters": [
                    {
                        "description": "User signup details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUpInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully created",
                        "schema": {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    },
                    "400": {
                        "description": "bas request",
                        "schema": {
                            "$ref": "#/definitions/models.ValidateErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: User with that email already exists",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway: Something bad happened",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateOrganizationInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "organization example 1"
                }
            }
        },
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Organization": {
            "type": "object",
            "properties": {
                "adminUser": {
                    "$ref": "#/definitions/models.User"
                },
                "adminUserID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Store"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.OrganizationRecord": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SignInInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "account1@mail.com"
                },
                "password": {
                    "type": "string",
                    "example": "Ruda,actv1!"
                }
            }
        },
        "models.SignInResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1N..."
                }
            }
        },
        "models.SignUpInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "passwordConfirm"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "account1@mail.com"
                },
                "name": {
                    "type": "string",
                    "example": "account1"
                },
                "password": {
                    "type": "string",
                    "example": "Ruda,actv1!"
                },
                "passwordConfirm": {
                    "type": "string",
                    "example": "Ruda,actv1!"
                }
            }
        },
        "models.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/models.Organization"
                },
                "organizationID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/models.Organization"
                },
                "organizationID": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "superAdmin": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "models.ValidateError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.ValidateErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ValidateError"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}